Make timeout configurable for async art uploads
The previously hardcoded timeout of 10 seconds is no
(always) sufficient for larger file uploads. Although
those files are uploaded correctly, the short timeout
causes the `content_id` to not be returned, which
hampers further processing.

This commit makes the `timeout` for the art uploads
configurable via a parameter, still defaulting
at the same 10 seconds.